main/xiaozhi-server/requirements.txt
# Core Dependencies
numpy==1.26.4  # Required for torch 2.2.2
torch==2.2.2
torchaudio==2.2.2

# Audio Processing
librosa~=0.8.1  # For voiceprint recognition
pyaudio==0.2.11  # For audio recording
silero_vad==5.1.2  # For voice activity detection
opuslib_next==1.1.2  # For opus audio format
pydub==0.25.1  # For audio file processing
funasr==1.2.3  # For speech recognition
edge_tts==7.0.0  # For text-to-speech
huggingface_hub==0.20.3  # For model downloading

# Network and API
websockets==14.2
httpx==0.27.2
aiohttp==3.9.3
aiohttp_cors==0.7.0
requests==2.32.3

# AI Services
openai==1.61.0
google-generativeai==0.8.4
mem0ai==0.1.62

# Voiceprint Recognition Tools
numba==0.52.0  # For audio processing optimization
tqdm==4.59.0  # For progress bars
visualdl==2.1.1  # For visualization
torchsummary~=1.5.1  # For model summary

# Configuration and Utils
pyyml==0.0.2
ormsgpack==1.7.0
ruamel.yaml==0.18.10
loguru==0.7.3
cozepy==0.12.0 